spiceflow-social/requirements.txt
beautifulsoup4==4.12.3
selectolax==0.3.21
httpx==0.27.2
aiohttp==3.9.5
async-timeout==4.0.3
python-dateutil==2.9.0
pytz==2024.1
icalendar==5.0.12
PyYAML==6.0.2
rich==13.8.0
# JS fallback (install later when allowed)
playwright==1.47.2
